Improve benchmark for measuring host-call performance.
Amortize the host calls across multiple invocations from Wasm, to reduce the
impact of overhead of calling into the Wasm module.

With this, it looks that it takes about 4-5ns per iteration of the following
loop
```
(loop $loop
        (if (i32.eq (i32.const 0) (local.get 0))
            (then (return (i32.const 0)))
            (else nop)
        )
        (call $sub_one (local.get 0))
        (local.set 0)
        (br $loop)
)
```

An additional benchmark of just calling into Wasm to execute the identity
function reveals that takes abtour 240ns, which fits with the measurements above.