Basic energy accounting with a host function
This adds a host function tick_energy(e: u32) that can be called from Wasm
to subtract the passed amount from the energy deposit. Execution is stopped
when the energy deposit is used up.

Currently this is initialized with a fixed example energy amount.

This will be refined in the future when he have a more clear specification for
our energy accounting.