Introduce artifacts and a compiler to them.
These are fully processed modules ready for execution. Execution of these
from byte form is intended to be cheap, specifically there will be no
deserialization, and the interepter will work on essentially the byte
representation.

Compared to Wasm code, artifacts have

- resolved jumps to absolute positions in the code sequence, removing relative
labels.
- remove the immediate arguments from the Operands, and instead encode them as
bytes. Whether this makes a difference is to be seen, but this way execution
easily operates on the byte array of instructions directly.