Artifact parser.
I made a choice to do a fairly low-level parser to (1) not copy data around
needlessly and (2) perhaps help with memory locality in execution, since all
code will be in one place. Likely this does not matter too much in the near
future, but I think the general principle is useful to keep in mind.

There are many more such low-level optimizations possible. For example, there is
no need for us to copy the data segments during parsing, we could just keep them
as-is.