Charge energy when advancing the iterator.
This cost is not constant and thus needs to be accounted for dynamically.

With more time we might be able to analyze and bound the cost, so that we could
be constant time per `next` call, but I do not have that insight now. However I
believe that should be possible.