Iterator locking with a separate trie.
This merges
https://github.com/Concordium/concordium-wasm-smart-contracts/pull/33 but
significantly revises it - remove the generation from the iterator. It is not
needed, iterators and entries are always implicitly for the latest generation. -
use a trie to keep track of locked prefixes. This makes it siginificantly more
efficient to check whether a part of a tree is locked. It has O(key length)
complexity, regardless of how many iterators or locks there are. - add
documentation - fix tests

Co-authored-by: Emil Lai <el@concordium.com>"