Revise node branching to only use 4 bits for branching.
The reason for the change is that the worst case space usage for updates is
lessened.

With branching on all 8 bits the amount of new data that needs to be stored when
updating a value with key length `n` is n * (2338 + C) where C is a small
constant that determines how much data is needed to store a pointer to a
value (so C is around 9, depending on whether the value is present or not).
The 2338 comes from the following calculation. For each node on the way to the
root of the tree we need to store
2 bytes for the length of the children list
256 * (1 + 9) (1 for the key, 9 for the pointer to the child)
32 bytes for the hash of the node

With branching on 4 bits the computation instead yields n * (354 + 2C).

With branching on 2 bits we have even less at n * (276 + 4C) however the
difference is much less, and using fewer bits increases both the amount of
indirection, amount of wasted space, and amount of computation (more hashes to
compute). Thus 4 bits seems like the most reasonable value.